requirements.txt
gradio
llama-cpp-python==0.2.75
requests